- Enlarged processor string identification size - Initial Brazos platform support (Ontario/Zacate). Be careful, it is totally untested at the moment! - Changed uname request in makefile (from -p to -m) - Switched processor interface FID and DID to float values instead of integers (brazos requires floating DIDs). Needs testing but seems to work ok...

+++ K10Processor.h
@@ -44,12 +44,12 @@ public:
 	void convertFreqtoFD(DWORD, int *, int *);
 		
 	void setVID (PState , DWORD);
-	void setFID (PState , DWORD);
-	void setDID (PState , DWORD);
+	void setFID (PState , float);
+	void setDID (PState , float);
 
 	DWORD getVID (PState);
-	DWORD getFID (PState);
-	DWORD getDID (PState);
+	float getFID (PState);
+	float getDID (PState);
 
 	void setFrequency (PState , DWORD);
 	void setVCore (PState, float);
